+++ src/Api/Deployments.php
@@ -18,13 +18,24 @@ class Deployments extends AbstractApi
 {
     /**
      * @param int|string $project_id
-     * @param array      $parameters
+     * @param array      $parameters {
+     *
+     *     @var string $order_by                    Return deployments ordered by id, iid, created_at, updated_at,
+     *                                              or ref fields (default is id)
+     *     @var string $sort                        Return deployments sorted in asc or desc order (default is desc)
+     * }
      *
      * @return mixed
      */
     public function all($project_id, array $parameters = [])
     {
         $resolver = $this->createOptionsResolver();
+        $resolver->setDefined('order_by')
+            ->setAllowedTypes('order_by', 'string')
+            ->setAllowedValues('order_by', ['id', 'iid', 'created_at', 'updated_at', 'ref']);
+        $resolver->setDefined('sort')
+            ->setAllowedTypes('sort', 'string')
+            ->setAllowedValues('sort', ['desc', 'asc']);
 
         return $this->get($this->getProjectPath($project_id, 'deployments'), $resolver->resolve($parameters));
     }