+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added CI schedule variables endpoints
 * Added support for triggering a pipeline
 * Added support for the search_namespaces projects parameter
+* Added support for order_by and sort deployments parameters
 
 [11.1.0]: https://github.com/GitLabPHP/Client/compare/11.0.0...11.1.0